requirements.txt
tensorflow-gpu==2.8.0
protobuf==3.20.0
pandas==1.3.5
scikit-learn==1.0.2
tqdm